update grouping interchangeable

+++ simjoin_entitymatching/value_matcher/fasttext.py
@@ -4,6 +4,7 @@ import gensim.models
 import joblib
 import pandas as pd
 import pathlib
+import os
 from collections import defaultdict
 from py_stringmatching.tokenizer.delimiter_tokenizer import DelimiterTokenizer
 
@@ -36,30 +37,24 @@ def _if_some_alphanumeric(s):
         if i.isalnum():
             return True
     return False
-    
-    
-def group_interchangeable_external_exp(target_attr, fasttext, default_match_res_dir="", 
-                                       default_icv_dir=""):
-    # apply
+
+
+def _infer_vec(tab, target_attr, fasttext):
     lattr = "ltable_" + target_attr
     rattr = "rtable_" + target_attr
     vec_dict = defaultdict()
     vec_pair = []
-
-    # read
-    partial_name, _ = ph.get_chunked_match_res_path(0, default_match_res_dir)   
-    partial_match_res = pd.read_csv(partial_name)
     
     # infer
     tokenizer = DelimiterTokenizer([' ', '\'', '\"', ',', '\\', '\t', '\r', '\n'])
     
-    for _, row in partial_match_res.iterrows():
+    for _, row in tab.iterrows():
         ori_lstr = row[lattr]
         ori_rstr = row[rattr]
         
         # check missing
         if pd.isnull(ori_lstr) or pd.isnull(ori_rstr):
-            raise ValueError(f"error in target attribute : {target_attr} with : {ori_lstr}, {ori_rstr}")
+            raise ValueError(f"error in target attribute : {lattr} or {rattr} with : {ori_lstr}, {ori_rstr}")
         
         # tokenize
         l_toks = tokenizer.tokenize(ori_lstr)
@@ -74,11 +69,11 @@ def group_interchangeable_external_exp(target_attr, fasttext, default_match_res_
         vec_dict[ori_lstr] = l_vecs
         vec_dict[ori_rstr] = r_vecs
         vec_pair.append((ori_lstr, ori_rstr))
+        
+    return vec_dict, vec_pair
 
-    # report
-    vec_path, pair_path = ph.get_icval_vec_input_path(_cur_parent_dir, default_icv_dir)
-    
-    # vec
+
+def _flush_vec(vec_dict, vec_path, vec_pair=[], pair_path="empty.pair"):
     with open(vec_path, "w") as vecfile:
         stat = [str(len(vec_dict)), '\n']
         vecfile.writelines(stat)
@@ -94,12 +89,32 @@ def group_interchangeable_external_exp(target_attr, fasttext, default_match_res_
                 vecfile.writelines(sub_v2)
             
     # candidare pair
-    with open(pair_path, "w") as pairfile:
-        stat = [str(len(vec_pair)), "\n"]
-        pairfile.writelines(stat)
-        for pair in vec_pair:
-            pairfile.writelines([pair[0] + "\n"])
-            pairfile.writelines([pair[1] + "\n"])
+    if os.path.exists(pair_path):
+        with open(pair_path, "w") as pairfile:
+            stat = [str(len(vec_pair)), "\n"]
+            pairfile.writelines(stat)
+            for pair in vec_pair:
+                pairfile.writelines([pair[0] + "\n"])
+                pairfile.writelines([pair[1] + "\n"])
             
-    return vec_dict
+    
+def group_interchangeable_external_exp(target_attr, fasttext, default_match_res_dir="", 
+                                       default_icv_dir=""):
+    # read
+    partial_name, neg_tab_bame = ph.get_chunked_match_res_path(0, default_match_res_dir)   
+    partial_match_res = pd.read_csv(partial_name)
+    neg_match_res = pd.read_csv(neg_tab_bame)
+    
+    # infer
+    pos_vec_dict, pos_vec_pair = _infer_vec(partial_match_res, target_attr, fasttext)
+    neg_vec_dict, _ = _infer_vec(neg_match_res, target_attr, fasttext)
+
+    # report
+    vec_path, pair_path = ph.get_icval_vec_input_path(_cur_parent_dir, default_icv_dir)
+    neg_vec_path, _ = ph.get_neg_icval_vec_input_path(_cur_parent_dir, default_icv_dir)
+    
+    _flush_vec(pos_vec_dict, pos_vec_pair, vec_path, pair_path)
+    _flush_vec(neg_vec_dict, neg_vec_path)
+            
+    return pos_vec_dict
